requirements.txt
beautifulsoup4==4.6.3
Flask==1.0.2
line-bot-sdk==1.8.0
requests==2.20.1
gunicorn